pron-server 会员获取邮件信息

+++ zlt-commons/zlt-common-core/src/main/java/com/central/common/model/KpnSiteAnnouncement.java
@@ -17,7 +17,7 @@ import lombok.EqualsAndHashCode;
 public class KpnSiteAnnouncement extends SuperEntity {
 
 	@ApiModelProperty(value = "站点id")
-	private int siteId;
+	private Long siteId;
 
 	@ApiModelProperty(value = "站点编码")
 	private String siteCode;
